% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Volcano and upset app}
\usage{
app(dataset)
}
\arguments{
\item{dataset}{A named list of datasets as output by maUEB::dea_toptable1()}
}
\description{
Runs a Shiny app for gene upsetting results exploration.
}
