% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{format_vol_data}
\alias{format_vol_data}
\title{Reformat volcano dataset for plotting
Expects as input the output of maUEB::dea_toptable1()}
\usage{
format_vol_data(
  dataset_list,
  log_fc_col = "logFC",
  unadj_p_col = "P.Value",
  adj_p_col = "adj.P.Val",
  extra_cols = c("Gene.Symbol", "EntrezID")
)
}
\arguments{
\item{dataset_list}{A named list of datasets where each entry contains the result of the comparison and is named after it.}

\item{log_fc_col}{The name of the column that contains the log fold value}

\item{unadj_p_col}{The name of the column that contains the unadjusted p value}

\item{adj_p_col}{The name of the column that contains the adjusted p value}

\item{tooltip_cols}{A character vector indicating which fields will be included in addition to the previous columns, normally used as tooltips in the graph}
}
\description{
Reformat volcano dataset for plotting
Expects as input the output of maUEB::dea_toptable1()
}
