% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{get_vol_spec}
\alias{get_vol_spec}
\title{Reformat volcano dataset for plotting
Expects as input the output of maUEB::dea_toptable1()}
\usage{
get_vol_spec(dataset, y = "adj_p", threshold = stats::rnorm(1))
}
\arguments{
\item{dataset}{a dataset from format_vol_vd}

\item{y}{the name of the column to use for y values}

\item{threshold}{value for the y axis thresholding. If used in a shiny app use the default value}
}
\description{
Reformat volcano dataset for plotting
Expects as input the output of maUEB::dea_toptable1()
}
