% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu_module}
\alias{menu_module}
\alias{menu_UI}
\alias{menu_server}
\title{Menu module}
\usage{
menu_UI(id, comparisons)

menu_server(id, db_time = 500)
}
\arguments{
\item{id}{Shiny id}

\item{comparisons}{a character vector with the comparisons in the dataset}

\item{db_time}{debounce time for the inputs}
}
\value{
A list with the following entries:
\itemize{
\item p: reactive numerical with the limit p value
\item adj: reactive logical indicating unadjusted or adjusted p value
\item comp: reactive character vector indicating the selected comparisons
\item log_fc_range: a reactive numerical vector with two entries \code{gt} and \code{lt}. When their corresponding checks are disabled they return \code{-Inf} and \code{Inf} respectively.
}
}
\description{
Provides inputs for the app
}
\section{Functions}{
\itemize{
\item \code{menu_UI()}: UI

\item \code{menu_server()}: UI

}}
