% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_module}
\alias{upset_module}
\alias{upset_UI}
\alias{upset_server}
\title{Upset module}
\usage{
upset_UI(id)

upset_server(id, dataset, p, adj_p, comp, log_fc_range)
}
\arguments{
\item{id}{Shiny ID}

\item{dataset}{Data as output by \code{format_vol_data}}

\item{p}{A reactive numerical indicating the p value limit}

\item{adj_p}{A reactive boolean indicating if adjusted p value should be used}

\item{log_fc_range}{A reactive numerical vector with two entries \code{lt} and \code{gt} indicating the limits for the Log FC}
}
\description{
Shows an upset chart
}
\section{Functions}{
\itemize{
\item \code{upset_UI()}: UI

\item \code{upset_server()}: server

}}
