% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_module}
\alias{volcano_module}
\alias{volcano_UI}
\alias{volcano_server}
\title{Volcano module}
\usage{
volcano_UI(id)

volcano_server(id, dataset, p, adj_p, comp, log_fc_range)
}
\arguments{
\item{id}{Shiny ID}

\item{dataset}{Data as output by \code{format_vol_data}}

\item{p}{A reactive numerical indicating the p value limit}

\item{adj_p}{A reactive boolean indicating if adjusted p value should be used}

\item{log_fc_range}{A reactive numerical vector with two entries \code{lt} and \code{gt} indicating the limits for the Log FC}
}
\value{
A list with one entry \code{chart} containing the UI of the module

A list with two entries:adj_p
\itemize{
\item \code{hovered}: Reactive containing a dataset with the hovered entry
\item \code{selected}: Reactive containing a dataset with the entries selected based on the p value and Log FC limits
}
}
\description{
Volcano module
}
\section{Functions}{
\itemize{
\item \code{volcano_UI()}: UI

\item \code{volcano_server()}: server

}}
